Refactoring process graphics code
